requirements.txt
streamlit
pandas
numpy
plotly
gspread
google-auth
gspread-dataframe
scikit-learn
xgboost